Refactor Attachment
Changed the order, `source` goes first and `type` is second. I just
thought that looked nicer.

I also introduce a protocol for URLStringConvertible so that you can
initialise the `Attachment` with a string instead of an `NSURL`. It
looks cleaner and works kind of in the same way.

To access the URL you just do `Attachment.source.url`.

What do you think @vadymmarkov ?